(feat): added the possibility of customizing id creation in adoption
